Update
Made nav bar consistent and functional, 'learn more' buttons on home work, started about page

Styling note: I added h1 tags to menu, locations, and about. The spacing between the navbar and the h1 is not the same on the menu page, which is because of the body margin-top being 65px on that page in particular.